Move function to the file that uses it
